+++ src/main/java/com/readum/infrastructure/aiChat/scheduler/SummarySchedulerConfig.java
@@ -1,4 +1,4 @@
-package com.readum.domain.aiChat.config;
+package com.readum.infrastructure.aiChat.scheduler;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
